split dependency.rs into multiple files and move document
