+++ src/project/mod.rs
@@ -1,0 +1,2 @@
+pub mod dependencies;
+pub mod document;
